refactor for better composition
