+++ src/app.js
@@ -1,0 +1,15 @@
+import React from 'react'
+import Header from './containers/header'
+import './app.css'
+
+
+export default function App() {
+  return (
+    <>
+      <Header/>
+      <article>
+        <h1>COUCOU</h1>
+      </article>
+    </>
+  )
+}
