--- src/components/Navbar/NavLink/navlink.css
@@ -1,27 +1,0 @@
-.nav-li {
-    margin-bottom: 1em;
-    text-align: center;
-}
-
-.nav-a {
-    text-decoration: none;
-    text-transform: uppercase;
-    font-size: 1.4rem;
-    color: white;
-    position: relative;
-}
-
-.nav-a::before {
-    content: '';
-    position: absolute;
-    bottom: 0;
-    width: 0%;
-    background: white;
-    height: 3px;
-    transition: width 100ms ease-in-out;
-}
-
-.nav-a:hover::before,
-.nav-a:focus::before {
-    width: 100%;
-}
